sdk/tests/fast/index-validation-expected.txt

Tests that index validation verifies the correct number of indices

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS gl.getError() is 0
PASS gl.drawElements(gl.TRIANGLES, 3, gl.UNSIGNED_SHORT, 0) is undefined.
PASS gl.getError() is 0
PASS successfullyParsed is true

TEST COMPLETE
